requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
discord==1.0.1
discord.py==1.5.1
expiringdict==1.2.1
idna==2.10
isort==5.6.4
kiwisolver==1.2.0
matplotlib==3.3.2
multidict==4.7.6
mypy-extensions==0.4.3
networkx==2.5
numpy==1.19.2
pandas==1.1.3
pathspec==0.8.0
Pillow==7.2.0
plotly==4.11.0
pptree==3.1
pycparser==2.20
pydot==1.4.1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.15
requests==2.24.0
retrying==1.3.3
scipy==1.5.2
six==1.15.0
text-to-image==0.0.5
texttable==1.6.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.0
